breaking solution in multiple scripts
would like that refactor ?
