+++ bioptim/optimization/receding_horizon_optimization.py
@@ -6,11 +6,11 @@ from time import perf_counter
 import numpy as np
 
 from .optimal_control_program import OptimalControlProgram
-from .solution import Solution
+from bioptim.optimization.solution.solution import Solution
 from ..dynamics.configure_problem import Dynamics, DynamicsList
 from ..limits.constraints import ConstraintFcn, ConstraintList
 from ..limits.objective_functions import ObjectiveFcn, ObjectiveList
-from ..limits.path_conditions import InitialGuess, InitialGuessList, BoundsList
+from ..limits.path_conditions import InitialGuessList, BoundsList
 from ..misc.enums import SolverType, InterpolationType, MultiCyclicCycleSolutions, ControlType
 from ..optimization.parameters import ParameterList
 from ..interfaces.solver_options import Solver
